Outsourced analysis code to a separate module.
Also, for the time being, there is always one forecast step for each
analysis step, instead of the number of forecast steps being changeable.
This is to make debugging easier.

I also fixed a bug caused by a list assignment that was presumed to
assign by value, but was actually assigning by reference.