Added observation perturbations to analysis scheme and increased ensemble size
Observation perturbation is essential when the integrator is deterministic; the ensemble
becomes too narrow otherwise and loses track of truth.

I also made better use of numpy arrays throughout the code. The ensemble
(and observation table) is now represented by a 2D numpy array, where
columns are the ensemble members.